Move PassingGradeConcepts to domain package
